requirements.txt
django-rest-framework==0.1.0
djangorestframework==3.17.1
asgiref==3.11.1
Django==6.0.2
psycopg2==2.9.11
psycopg2-binary==2.9.11
sqlparse==0.5.5
tzdata==2025.3
django-filter==25.2